Move findM to shared/catsContrib
